extract mocks + utility functions in external files
