+++ src/internal/test/data-testid-helpers.ts
@@ -1,0 +1,3 @@
+export const queryByTestId = <T extends Element>(container: HTMLElement, testId: string): T | null => {
+  return container.querySelector<T>(`[data-testid=${testId}]`);
+};
